+++ app/urls.py
@@ -4,9 +4,4 @@ from .views import *
 
 urlpatterns = [
     path('', IndexView.as_view(), name='index'),
-    path('login', LoginView.as_view(), name='login'),
-    path('logout', LogoutView.as_view(), name='logout'),
-    path('register', RegisterView.as_view(), name='register'),
-    path('seats', SeatsView.as_view(), name='seats'),
-    path('attend', AttendanceView.as_view(), name='attend')
 ]
