+++ conf/urls.py
@@ -3,5 +3,7 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('app.urls'))
+    path('', include('app.urls')),
+    path('', include('user.urls')),
+    path('seats', include('seats.urls'))
 ]
