+++ seats/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+from .models import ClassModel, SeatModel
+
+admin.site.register(ClassModel)
+admin.site.register(SeatModel)
+
+# Register your models here.
