+++ templates/seats.html
@@ -54,7 +54,7 @@
             <button>출석관리</button>
           </form>
         {% else %}
-          <form class="check" action="/attend" method="post">
+          <form class="check" action="/seats/attend" method="post">
             <button>출석하기</button>
           </form>
         {% endif %}
